+++ src/finalisers/es6.js
@@ -1,6 +1,16 @@
-import { keys } from '../utils/object';
+import { blank, keys } from '../utils/object';
 
-export default function es6 ( bundle, magicString, { exportMode }, options ) {
+function uniqueNames ( declarations ) {
+	let uniques = blank();
+
+	declarations
+		.filter( declaration => !/^(default|\*)$/.test( declaration.name ) )
+		.forEach( declaration => uniques[ declaration.name ] = true );
+
+	return keys( uniques );
+}
+
+export default function es6 ( bundle, magicString ) {
 	const importBlock = bundle.externalModules
 		.map( module => {
 			const specifiers = [];
@@ -16,10 +26,7 @@ export default function es6 ( bundle, magicString, { exportMode }, options ) {
 			}
 
 			if ( module.needsNamed ) {
-				specifiers.push( '{ ' + module.importedByBundle
-					.filter( declaration => !/^(default|\*)$/.test( declaration.name ) )
-					.map( ({ name, localName }) =>
-						name === localName ? name : `${name} as ${localName}` )
+				specifiers.push( '{ ' + uniqueNames( module.importedByBundle )
 					.join( ', ' ) + ' }' );
 			}
 
@@ -37,7 +44,7 @@ export default function es6 ( bundle, magicString, { exportMode }, options ) {
 	const exportBlock = keys( exports ).map( exportedName => {
 		const specifier = exports[ exportedName ];
 
-		const canonicalName = bundle.entryModule.getCanonicalName( specifier.localName );
+		const canonicalName = bundle.entryModule.replacements[ specifier.localName ] || specifier.localName;
 
 		if ( exportedName === 'default' ) {
 			return `export default ${canonicalName};`;