+++ src/finalisers/shared/getExportBlock.js
@@ -1,12 +1,18 @@
 export default function getExportBlock ( bundle, exportMode, mechanism = 'return' ) {
 	if ( exportMode === 'default' ) {
-		return `${mechanism} ${bundle.entryModule.getCanonicalName('default')};`;
+		const defaultExport = bundle.entryModule.exports.default;
+
+		const defaultExportName = bundle.entryModule.replacements.default ||
+			defaultExport.identifier;
+
+		return `${mechanism} ${defaultExportName};`;
 	}
 
 	return bundle.toExport
 		.map( name => {
 			const prop = name === 'default' ? `['default']` : `.${name}`;
-			return `exports${prop} = ${bundle.entryModule.getCanonicalName(name)};`;
+			name = bundle.trace( bundle.entryModule, name );
+			return `exports${prop} = ${name};`;
 		})
 		.join( '\n' );
 }