+++ test/form/dedupes-external-imports/_config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	description: 'dedupes external imports',
+	options: {
+		external: [ 'external' ]
+	}
+};