+++ test/form/dedupes-external-imports/bar.js
@@ -1,0 +1,8 @@
+import { Component } from 'external';
+
+export default class Bar extends Component {
+	constructor () {
+		super();
+		this.isBar = true;
+	}
+}