+++ test/form/dedupes-external-imports/main.js
@@ -1,0 +1,7 @@
+import Foo from './foo';
+import Bar from './bar';
+import Baz from './baz';
+
+const foo = new Foo();
+const bar = new Bar();
+const baz = new Baz();