+++ test/form/unused-default-exports/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'does not name unused-but-included default export'
+};