+++ test/function/deconflicts-generated-default-names/foo.js
@@ -1,0 +1,9 @@
+export default notActuallyFoo;
+
+function notActuallyFoo () {
+  return 'not ' + foo();
+}
+
+function foo () {
+  return 'actually foo';
+}