+++ test/function/deconflicts-generated-default-names/main.js
@@ -1,0 +1,3 @@
+import foo from './foo';
+
+assert.equal( foo(), 'not actually foo' );