+++ test/function/export-default-as-b/main.js
@@ -1,0 +1,3 @@
+import { baz } from './bar';
+
+assert.equal( baz, 'BAZ' );