+++ test/function/export-default-as-c/main.js
@@ -1,0 +1,5 @@
+import * as namespace from './namespace';
+
+assert.equal( namespace.baz, 'BAZ' );
+
+export { namespace };