+++ test/function/export-default-as-c/namespace.js
@@ -1,0 +1,1 @@
+export { default as baz } from './baz';