+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # rollup changelog
 
+## 0.14.0
+
+* Internal refactoring
+* Correctly deconflict generated default export names ([#72](https://github.com/rollup/rollup/issues/72))
+* Handle `export { x } from 'y'` declarations ([#74](https://github.com/rollup/rollup/issues/74))
+* Dedupe named imports from external modules in ES6 bundles ([#77](https://github.com/rollup/rollup/issues/77))
+
 ## 0.13.0
 
 * Support `banner` and `footer` options ([#66](https://github.com/rollup/rollup/pull/66))